@@ -4,7 +4,6 @@ import (
 	"database/sql"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"net"
 	"os"
 	"os/user"
@@ -107,7 +106,7 @@ func Test_ErrorWhenUnableToInitDatabase(t *testing.T) {
 	jarFile, cleanUp := createTempXzArchive()
 	defer cleanUp()
 
-	extractPath, err := ioutil.TempDir(filepath.Dir(jarFile), "extract")
+	extractPath, err := os.MkdirTemp(filepath.Dir(jarFile), "extract")
 	if err != nil {
 		panic(err)
 	}
@@ -143,7 +142,7 @@ func Test_ErrorWhenUnableToCreateDatabase(t *testing.T) {
 
 	defer cleanUp()
 
-	extractPath, err := ioutil.TempDir(filepath.Dir(jarFile), "extract")
+	extractPath, err := os.MkdirTemp(filepath.Dir(jarFile), "extract")
 
 	if err != nil {
 		panic(err)
@@ -214,7 +213,7 @@ func Test_ErrorWhenCannotStartPostgresProcess(t *testing.T) {
 
 	defer cleanUp()
 
-	extractPath, err := ioutil.TempDir(filepath.Dir(jarFile), "extract")
+	extractPath, err := os.MkdirTemp(filepath.Dir(jarFile), "extract")
 	if err != nil {
 		panic(err)
 	}
@@ -236,7 +235,7 @@ func Test_ErrorWhenCannotStartPostgresProcess(t *testing.T) {
 }
 
 func Test_CustomConfig(t *testing.T) {
-	tempDir, err := ioutil.TempDir("", "embedded_postgres_test")
+	tempDir, err := os.MkdirTemp("", "embedded_postgres_test")
 	if err != nil {
 		panic(err)
 	}
@@ -281,7 +280,7 @@ func Test_CustomConfig(t *testing.T) {
 }
 
 func Test_CustomLog(t *testing.T) {
-	tempDir, err := ioutil.TempDir("", "embedded_postgres_test")
+	tempDir, err := os.MkdirTemp("", "embedded_postgres_test")
 	if err != nil {
 		panic(err)
 	}
@@ -402,7 +401,7 @@ func Test_CanStartAndStopTwice(t *testing.T) {
 }
 
 func Test_ReuseData(t *testing.T) {
-	tempDir, err := ioutil.TempDir("", "embedded_postgres_test")
+	tempDir, err := os.MkdirTemp("", "embedded_postgres_test")
 	if err != nil {
 		panic(err)
 	}
@@ -480,7 +479,7 @@ func Test_ReuseData(t *testing.T) {
 }
 
 func Test_CustomBinariesRepo(t *testing.T) {
-	tempDir, err := ioutil.TempDir("", "embedded_postgres_test")
+	tempDir, err := os.MkdirTemp("", "embedded_postgres_test")
 	if err != nil {
 		panic(err)
 	}
@@ -526,7 +525,7 @@ func Test_CustomBinariesRepo(t *testing.T) {
 }
 
 func Test_CustomBinariesLocation(t *testing.T) {
-	tempDir, err := ioutil.TempDir("", "prepare_database_test")
+	tempDir, err := os.MkdirTemp("", "prepare_database_test")
 	if err != nil {
 		panic(err)
 	}
@@ -564,12 +563,12 @@ func Test_CustomBinariesLocation(t *testing.T) {
 }
 
 func Test_PrefetchedBinaries(t *testing.T) {
-	binTempDir, err := ioutil.TempDir("", "prepare_database_test_bin")
+	binTempDir, err := os.MkdirTemp("", "prepare_database_test_bin")
 	if err != nil {
 		panic(err)
 	}
 
-	runtimeTempDir, err := ioutil.TempDir("", "prepare_database_test_runtime")
+	runtimeTempDir, err := os.MkdirTemp("", "prepare_database_test_runtime")
 	if err != nil {
 		panic(err)
 	}
@@ -616,7 +615,7 @@ func Test_PrefetchedBinaries(t *testing.T) {
 }
 
 func Test_RunningInParallel(t *testing.T) {
-	tempPath, err := ioutil.TempDir("", "parallel_tests_path")
+	tempPath, err := os.MkdirTemp("", "parallel_tests_path")
 	if err != nil {
 		panic(err)
 	}
